# cars-service/src/main/java/za/co/protogen/core/domain/api/UpdateApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package za.co.protogen.core.domain.api;

import za.co.protogen.core.domain.models.delegate.UpdateApiDelegate;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-11-07T20:54:30.516515700+02:00[Africa/Johannesburg]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "car", description = "the car API")
public interface UpdateApi {

    default UpdateApiDelegate getDelegate() {
        return new UpdateApiDelegate() {};
    }

    /**
     * PUT /update/{vin}/{sWhat}/{sTo} : Update car
     *
     * @param vin Car vin (Vehicle Identification Number) used to retrieve car (required)
     * @param sWhat The letter corresponding to the field to Update (required)
     * @param sTo new data to replace the data of a chosen field with (required)
     * @return Successful Responds. Return message and updated car (status code 200)
     *         or Letter that is not an option selected (status code 400)
     */
    @Operation(
        operationId = "updateCar",
        summary = "Update car",
        tags = { "car", "update" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful Responds. Return message and updated car", content = {
                @Content(mediaType = "text/html", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "400", description = "Letter that is not an option selected", content = {
                @Content(mediaType = "text/html", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/update/{vin}/{sWhat}/{sTo}",
        produces = { "text/html" }
    )
    
    default ResponseEntity<String> updateCar(
        @Parameter(name = "vin", description = "Car vin (Vehicle Identification Number) used to retrieve car", required = true, in = ParameterIn.PATH) @PathVariable("vin") String vin,
        @Parameter(name = "sWhat", description = "The letter corresponding to the field to Update", required = true, in = ParameterIn.PATH) @PathVariable("sWhat") String sWhat,
        @Parameter(name = "sTo", description = "new data to replace the data of a chosen field with", required = true, in = ParameterIn.PATH) @PathVariable("sTo") String sTo
    ) {
        return getDelegate().updateCar(vin, sWhat, sTo);
    }

}